requirements.txt
# https://raw.githubusercontent.com/apache/airflow/constraints-2.9.2/constraints-3.8.txt
pandas
typing-extensions==4.12.2
pydantic==2.7.4
psycopg2-binary==2.9.9
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
SQLAlchemy==1.4.52
apache-airflow
apache-airflow-providers-jdbc==4.3.1
apache-airflow-providers-microsoft-mssql==3.7.1
apache-airflow-providers-postgres==5.11.1
apache-airflow-providers-tableau==4.5.1
python-daemon==3.0.1
python-dotenv==1.0.1
python-telegram-bot==20.2
python-http-client==3.3.7
regex==2024.5.15
requests-file==2.1.0
requests==2.31.0
responses==0.25.0
tableauserverclient==0.25
kubernetes==29.0.0
kubernetes_asyncio==29.0.0
types-toml==0.10.8.20240310











